# planner: fix a panic of a cached prepared statement with IndexScan