@dbjoa 2018-10-23
/run-all-tests