@dbjoa 2018-10-25
@winoros I've updated the PR.