@dbjoa 2018-10-25
/run-all-tests